# apps/api/src/waitlist/dto/join-waitlist.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsEmail, IsNotEmpty } from 'class-validator';

export class JoinWaitlistDto {
  @ApiProperty({
    description: 'Email address to join the waitlist',
    example: 'user@example.com',
  })
  @IsEmail({}, { message: 'Please provide a valid email address' })
  @IsNotEmpty({ message: 'Email is required' })
  email: string;
}