# packages/api-client/src/models/JoinWaitlistDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type JoinWaitlistDto = {
    /**
     * Email address to join the waitlist
     */
    email: string;
};