# packages/api-client/src/models/VerifyResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VerifyResponseDto = {
    /**
     * Success message
     */
    message: string;
    /**
     * Indicates if user was added to Resend
     */
    addedToResend: boolean;
};