# packages/api-client/src/models/WaitlistResponseDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type WaitlistResponseDto = {
    /**
     * Success message
     */
    message: string;
    /**
     * Email address submitted
     */
    email: string;
};